refactor(type): Move election functions to election_utils file
Need this to avoid a dependency cycle in upcoming commits
